Child component Under Respective Parent & Perf++
