+++ src/app/app-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { CurrentseasonComponent } from './currentseason/currentseason.component';
+import { CurrentseasonComponent } from './searchbox/currentseason/currentseason.component';
 import { EpisodesComponent } from './episodes/episodes.component';
 import { HomeComponent } from './home/home.component';
-import { RecommendationsComponent } from './recommendations/recommendations.component';
-import { SearchResultComponent } from './search-result/search-result.component';
-import { TopComponent } from './top/top.component';
-import { UpcomingComponent } from './upcoming/upcoming.component';
+import { RecommendationsComponent } from './episodes/recommendations/recommendations.component';
+import { SearchResultComponent } from './searchbox/search-result/search-result.component';
+import { TopComponent } from './searchbox/top/top.component';
+import { UpcomingComponent } from './searchbox/upcoming/upcoming.component';
 
 const routes: Routes = [
   {path:"",component:HomeComponent},
