+++ src/app/app.component.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
 
 <head>
      <meta charset="utf-8">
