+++ src/app/app.module.ts
@@ -3,25 +3,24 @@ import { BrowserModule } from '@angular/platform-browser';
 import { HttpClientModule } from '@angular/common/http';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { SearchResultComponent } from './search-result/search-result.component';
+import { SearchResultComponent } from './searchbox/search-result/search-result.component';
 import { SearchboxComponent } from './searchbox/searchbox.component';
 import { FormsModule } from '@angular/forms';
 import { EpisodesComponent } from './episodes/episodes.component';
 import { HrtominPipe } from './pipes/hrtomin.pipe';
-import { RecommendationsComponent } from './recommendations/recommendations.component';
+import { RecommendationsComponent } from './episodes/recommendations/recommendations.component';
 import { HomeComponent } from './home/home.component';
-import { TopComponent } from './top/top.component';
-import { ReviewsComponent } from './reviews/reviews.component';
-import { RelationsComponent } from './relations/relations.component';
-import { ImagesComponent } from './images/images.component';
-import { CurrentseasonComponent } from './currentseason/currentseason.component';
-import { UpcomingComponent } from './upcoming/upcoming.component';
-import { RandomComponent } from './random/random.component';
+import { TopComponent } from './searchbox/top/top.component';
+import { ReviewsComponent } from './episodes/reviews/reviews.component';
+import { RelationsComponent } from './episodes/relations/relations.component';
+import { ImagesComponent } from './episodes/images/images.component';
+import { CurrentseasonComponent } from './searchbox/currentseason/currentseason.component';
+import { UpcomingComponent } from './searchbox/upcoming/upcoming.component';
+import { RandomComponent } from './home/random/random.component';
 import { TotopComponent } from './totop/totop.component';
 import { LoadingSpinnerComponent } from './loading-spinner/loading-spinner.component';
-import { EpisodevidsComponent } from './episodevids/episodevids.component';
-import { VideoplayerComponent } from './videoplayer/videoplayer.component';
-
+import { EpisodevidsComponent } from './episodes/episodevids/episodevids.component';
+//videogular
 import { VgCoreModule } from 'ngx-videogular';
 import { VgControlsModule } from 'ngx-videogular';
 import { VgOverlayPlayModule } from 'ngx-videogular';
@@ -50,13 +49,14 @@ import { VgStreamingModule } from 'ngx-videogular';
     TotopComponent,
     LoadingSpinnerComponent,
     EpisodevidsComponent,
-    VideoplayerComponent,
+ 
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
     HttpClientModule,
     FormsModule,
+    //videogular
     VgCoreModule,
     VgControlsModule,
     VgOverlayPlayModule,
