+++ src/app/episodes/recommendations/recommendations.component.css
@@ -1,0 +1,25 @@
+.containerR {
+  margin-top: 0px;
+  color: #1c87c9;
+  height:auto;
+  overflow: hidden;
+  width: auto;
+  padding-left: 0px;
+  padding-right: 0px;
+  break-inside: avoid;
+  display: block;
+  justify-content: center;
+  padding-bottom: 20px;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+/* heading */
+.h1 {
+  font-family: 'Alexandria', sans-serif;
+  font-size: 20px;
+  padding-left: 10px;
+  padding-top: 10px;
+  padding-bottom: 0px;
+
+}
