+++ src/app/home/random/random.component.css
