+++ src/app/home/random/random.component.html
@@ -9,7 +9,7 @@
 
     <!-- <--------------- Main content-------------->
     <div class="both" *ngIf="dataload">
-        <div class="left" ><img class="img" [src]="datar.data.images.jpg.image_url"></div>
+        <div class="left" ><img alt="anime"class="img" [src]="datar.data.images.jpg.image_url"></div>
         <div class="left" >
             <ul class="list-group list-group-flush" style="text-align:left;" *ngIf="dataload">
                 <li class="list-group-item" *ngIf="datar.data.title "><a class="title">Title:</a> <a
