--- src/app/recommendations/recommendations.component.css
@@ -1,230 +1,0 @@
-.containerR {
-  margin-top: 0px;
-  /* background: black; */
-  color: #1c87c9;
-
-  height:auto;
-  overflow: hidden;
-  width: auto;
-  padding-left: 0px;
-  padding-right: 0px;
-  /* display: table; */
-  break-inside: avoid;
-  display: block;
-
-  justify-content: center;
-  padding-bottom: 20px;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-/* line */
-.hr1 {
-  border: 0;
-  clear: both;
-  display: block;
-  width: 100%;
-  background-color: #d6d632;
-  height: 1px;
-}
-
-/* heading */
-.h1 {
-  font-family: 'Alexandria', sans-serif;
-  font-size: 20px;
-  padding-left: 10px;
-  padding-top: 10px;
-  padding-bottom: 0px;
-
-}
-
-@media screen and (max-width: 767px) {
-  .container {
-    min-height: 50vh;
-    padding-bottom: 5px;
-    margin-top: 5px;
-    margin-bottom: 5px;
-
-  }
-}
-
-
-/* data */
-
-.heading {
-  font-size: 20px;
-  font-family: 'Zen Kurenaido', sans-serif;
-  font-weight: bold;
-
-}
-
-.viewAll {
-  font-family: 'Mulish', sans-serif;
-  position: relative;
-  left: 35%;
-  font-weight: bold;
-  font-size: 1.1em;
-  cursor: pointer;
-  width: 82px;
-
-}
-
-
-.prevnex:hover {
-  transform: scale(1.1);
-  transition: opacity .9s,transform .5s,transform .5s;
-  -webkit-transition: opacity .9s, transform .5s, transform .5s;
-}
-
-.Image-Name:hover {
-  color: #1c87c9;
-  transform: scale(1.1);
-  transition: opacity .9s,transform .5s,transform .5s;
-  -webkit-transition: opacity .9s, transform .5s, transform .5s;
-
-}
-
-.Showing {
-  padding: 5px;
-}
-
-.Image-Name:hover .show-name {
-  white-space: initial;
-  text-overflow: initial;
-  word-wrap: normal;
-  width: 210px;
-  display: block;
-  overflow: visible;
-  transform: scale(1.1);
-  transition: opacity .9s,transform .5s,transform .5s;
-  -webkit-transition: opacity .9s, transform .5s, transform .5s;
-
-}
-
-#Banners-container {
-  margin-top: 5px;
-  width: 217px;
-  margin-left: auto;
-  margin-right: auto;
-
-  flex-wrap: wrap;
-  display: flex;
-  border-color: aliceblue;
-  margin-left: 10px;
-}
-
-#Banners-container img {
-  width: 217px;
-  border-radius: 5px;
-  margin-right: 20px;
-  margin-top: 0px;
-  cursor: pointer;
-}
-
-.Image-Name {
-  position: relative;
-  color: white;
-  width: 150px;
-
-
-}
-
-.ep {
-  background: red;
-  color: white;
-  padding: 2px 12px;
-  border-radius: 3px;
-  position: absolute;
-  display: inline-block;
-  top: 25px;
-  left: 10px;
-  font-size: 0.9em;
-  font-family: 'Ubuntu', sans-serif;
-}
-
-.epair {
-  background: #8ebf42;
-  color: black;
-  padding: 2px 12px;
-  border-radius: 3px;
-  position: absolute;
-  display: inline-block;
-  top: 25px;
-  left: 10px;
-  font-size: 0.9em;
-  font: 900;
-  font-family: 'Ubuntu', sans-serif;
-}
-
-.comments {
-  background: rgb(32, 241, 216);
-  padding: 2px 8px;
-  border-radius: 3px;
-  color: black;
-  font-weight: bold;
-  position: absolute;
-  top: 295px;
-  left: 10px;
-  font-size: 0.5em;
-  font-family: 'Ubuntu', sans-serif;
-}
-
-
-
-.watch {
-  background: rgb(32, 241, 216);
-  padding: 2px 8px;
-  position: absolute;
-  margin-right: 4px;
-  margin-bottom: 4px;
-  top: 295px;
-  left: 150px;
-  border-radius: 3px;
-  color: black;
-  font-weight: bold;
-  font-size: 0.5em;
-  font-family: 'Ubuntu', sans-serif;
-}
-
-.content-type {
-  display: flex;
-  height: 13px;
-  margin: 7px 0 7px 0;
-}
-
-.action {
-  display: flex;
-  background: rgb(49, 49, 49);
-  border-radius: 17px;
-  padding: 6px 3px;
-  font-size: 10px;
-  align-items: center;
-}
-
-.show-name {
-  font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-  font-size: 15px;
-  cursor: default;
-  margin-bottom: 5px;
-  white-space: pre-wrap;
-  text-overflow: ellipsis;
-  word-wrap: normal;
-  width: 210px;
-  overflow: visible;
-  font-weight: 600;
-}
-col{
-  padding: 10px;
-
-}
-.epcon {
-  display: contents;
-  margin-top: 20px;
-
-}
-.row {
-  --bs-gutter-x: 0;
-  --bs-gutter-y: 0;
-  display: flex;
-  flex-wrap: wrap;
-}
