+++ src/app/searchbox/search-result/search-result.component.css
