+++ src/app/searchbox/top/top.component.css
@@ -1,0 +1,19 @@
+.container {
+  border-radius: 20px;
+
+  margin-top: 15px;
+  background: black;
+  color: #1c87c9;
+  min-height: 70vh;
+  width: 90vw;
+  display: table;
+  padding-bottom: 20px;
+
+
+
+}
+
+
+
+
+
