+++ src/app/searchbox/upcoming/upcoming.component.css
@@ -1,0 +1,26 @@
+.container {
+  border-radius: 20px;
+  margin-top: 15px;
+  background: black;
+  color: #1c87c9;
+  min-height: 70vh;
+  width: 90vw;
+  display: table;
+  padding-bottom: 20px;
+
+
+
+}
+
+@media screen and (max-width: 767px) {
+  .container {
+    min-height: 50vh;
+    padding-bottom: 5px;
+    margin-top: 5px;
+    margin-bottom: 5px;
+
+  }
+}
+
+
+/* data */
