+++ src/app/searchbox/upcoming/upcoming.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { AnyForUntypedForms } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
-import { SearchResultServiceService } from '../services/search-result-service.service';
+import { SearchResultServiceService } from '../../services/search-result-service.service';
 declare var window: any;
