--- src/assets/custom.js
@@ -1,16 +1,0 @@
-function downloadImage(url, name){
-    fetch(url)
-      .then(resp => resp.blob())
-      .then(blob => {
-          const url = window.URL.createObjectURL(blob);
-          const a = document.createElement('a');
-          a.style.display = 'none';
-          a.href = url;
-          // the filename you want
-          a.download = name;
-          document.body.appendChild(a);
-          a.click();
-          window.URL.revokeObjectURL(url);
-      })
-      .catch(() => alert('An error sorry'));
-}
