+++ src/styles.css
@@ -10,8 +10,6 @@
     opacity: .6;
 }
 
-
-
 @keyframes spin {
     0% { transform: rotate(0deg); }
     100% { transform: rotate(360deg); }
@@ -19,3 +17,192 @@
 body{
     background-color: black !important;
 }
+
+h1{
+    font-size: 20px;
+  }
+  .hr1{
+    border: 0;
+    clear: both;
+    display: block;
+    width: 100%;
+    background-color: #FFFF00;
+    height: 1px;
+  }
+
+/* 
+  data */
+
+  
+
+.heading {
+    font-size: 20px;
+    font-family: 'Zen Kurenaido', sans-serif;
+    font-weight: bold;
+  
+  }
+  
+  .viewAll {
+    font-family: 'Mulish', sans-serif;
+    position: relative;
+    left: 35%;
+    font-weight: bold;
+    font-size: 1.1em;
+    cursor: pointer;
+    width: 82px;
+  
+  }
+  
+  
+  .prevnex:hover {
+    color: #8ebf42
+  
+  }
+  
+  .Image-Name:hover {
+    color: #1c87c9;
+    transform: scale(1.1);
+    transition: opacity .9s,transform .5s,transform .5s;
+    -webkit-transition: opacity .9s, transform .5s, transform .5s;
+  
+  }
+  
+  .Showing {
+    padding: 5px;
+  }
+  
+  .Image-Name:hover .show-name {
+  
+    white-space: initial;
+    text-overflow: initial;
+    word-wrap: normal;
+    width: 210px;
+    display: block;
+    overflow: visible;
+    transform: scale(1.1);
+    transition: opacity .9s,transform .5s,transform .5s;
+    -webkit-transition: opacity .9s, transform .5s, transform .5s;
+  
+  }
+  
+  #Banners-container {
+    margin-top: 5px;
+    width: 217px;
+  
+    flex-wrap: wrap;
+    display: flex;
+    border-color: aliceblue;
+    margin-left: 15px;
+  }
+  
+  #Banners-container img {
+    width: 217px;
+    border-radius: 5px;
+    margin-right: 20px;
+    margin-top: 12px;
+    cursor: pointer;
+  }
+  
+  .Image-Name {
+    position: relative;
+    color: white;
+    width: 217px;
+  
+  
+  }
+  
+  .ep {
+    background: red;
+    color: white;
+    padding: 2px 12px;
+    border-radius: 3px;
+    position: absolute;
+    display: inline-block;
+    top: 25px;
+    left: 10px;
+    font-size: 0.9em;
+    font-family: 'Ubuntu', sans-serif;
+  }
+  
+  .epair {
+    background: #8ebf42;
+    color: black;
+    padding: 2px 12px;
+    border-radius: 3px;
+    position: absolute;
+    display: inline-block;
+    top: 25px;
+    left: 10px;
+    font-size: 0.9em;
+    font: 900;
+    font-family: 'Ubuntu', sans-serif;
+  }
+  
+  .rating {
+    background: rgb(32, 241, 216);
+    padding: 2px 8px;
+    border-radius: 3px;
+    color: black;
+    font-weight: bold;
+    position: absolute;
+    top: 295px;
+    left: 10px;
+    font-size: 0.5em;
+    font-family: 'Ubuntu', sans-serif;
+  }
+  
+  
+  
+  .watch {
+    background: rgb(32, 241, 216);
+    padding: 2px 8px;
+    position: absolute;
+    margin-right: 4px;
+    margin-bottom: 4px;
+    top: 295px;
+    left: 150px;
+    border-radius: 3px;
+    color: black;
+    font-weight: bold;
+    font-size: 0.5em;
+    font-family: 'Ubuntu', sans-serif;
+  }
+  
+  .content-type {
+    display: flex;
+    height: 13px;
+    margin: 7px 0 7px 0;
+  }
+  
+  
+  .type {
+    display: flex;
+    background: rgb(49, 49, 49);
+    border-radius: 17px;
+    padding: 6px 3px;
+    font-size: 10px;
+    align-items: center;
+  }
+  
+  
+  
+  .show-name {
+  
+    /* font-family: 'Nunito', sans-serif; */
+    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+    font-size: 15px;
+    cursor: default;
+    margin-bottom: 5px;
+    white-space: pre-wrap;
+    text-overflow: ellipsis;
+    word-wrap: normal;
+    width: 210px;
+    overflow: visible;
+    font-weight: 600;
+  }
+  
+  .epcon {
+    display: contents;
+    margin-top: 20px;
+  
+  }
